Variable Definition and Interface Complete, GUI 40% done in total(estimate)

+++ ES2-2018-EIC2-03/src/optimization/DataType.java
@@ -6,5 +6,5 @@ public enum DataType {
 	//e nos casos em que seja possível, o intervalo de valores aceitável 
 	//para as variáveis de decisão (por exemplo, para variáveis do tipo real, [-5;5]);
 	/*Comentário Pessoal: Que outros tipos de dados podemos ter?! O que é aceite pelo jMetal? Em espera*/
-	INTEGER, DOUBLE, BINARY;
+	INTEGER, REAL, BINARY;
 }
